14-advanced-retrieval/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Vector Database
langchain-chroma>=0.1.0
chromadb>=0.4.0

# BM25 Retrieval (requires rank-bm25)
rank-bm25>=0.2.2

# For environment variable management
python-dotenv>=1.0.0
